whole rearrangement of test_feas in MCND_lpcheckfeas and bug fix in flowvar logical bound reductions --- v7.6.0

+++ mcnd/cg/globalcutmanager.cpp
@@ -77,7 +77,7 @@ GlobalCutManager::clean_collection(){
 //------------------------------------------------------------------------------------------
 
 int
-GlobalCutManager::globalc_generation_main( const double * ystar, const double * closed, int cont0, int curr_id){
+GlobalCutManager::globalc_generation_main( const double * ystar, const int * closed, int cont0, int curr_id){
     int opnd=0;
     int size;
     int * vars_;
