+++ mcnd/include/globalcutmanager.hpp
@@ -38,7 +38,7 @@ public:
     //-------------------------------------------------------------------------------------------
     //  main methods
     //-------------------------------------------------------------------------------------------
-    int globalc_generation_main( const double * ystar, const double * closed, int cont0, int curr_id);
+    int globalc_generation_main( const double * ystar, const int * closed, int cont0, int curr_id);
     int make_globalcut(const double * ystar, int sz,  int* vars_,  int curr_id);
      
     //-------------------------------------------------------------------------------------------
