+++ mcnd/solver/OsiVolSolverInterface.cpp
@@ -872,7 +872,7 @@ OsiVolSolverInterface::add_external_localc( const int * y, const double * sol, i
 //-----------------------------------------------------------------------
 
 int 
-OsiVolSolverInterface::add_external_globalc( const double * y, int cont0){
+OsiVolSolverInterface::add_external_globalc( const int * y, int cont0){
 	
      
  	int ret = globalc_manager->globalc_generation_main(solution , y, cont0, numrows_);
